fix HEAD exception error
Ignore-this: 350f1886464f4aa5719359b32917ac32
at the base exception handler for a client connection we assumed that the exception
was a condition, whereas it is not always one.
This led to another fatal exception.
Maybe it should be, but a special record type is used for certain errors.
We should use the guard form's predicates to handle this in the future.
But for now hack it so that HEAD doesn't totally kill the server.

darcs-hash:20110504135145-ebcb5-72398c75ff6bf65b53ee500370e7bc0c65be1fe7.gz